Move config.
