+++ crates/core/config/src/tasks/config.rs
@@ -6,7 +6,7 @@ use crate::errors::{
 use crate::helpers::gather_extended_sources;
 use crate::project::TaskConfig;
 use crate::types::FileGroups;
-use crate::validators::{is_default, validate_extends, validate_id};
+use crate::validators::{is_default, validate_extends, validate_id, validate_target};
 use figment::{
     providers::{Format, YamlExtended},
     Figment,
@@ -16,6 +16,21 @@ use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use std::path::PathBuf;
 use validator::{Validate, ValidationError};
+
+fn validate_deps(list: &[String]) -> Result<(), ValidationError> {
+    for (index, item) in list.iter().enumerate() {
+        let key = format!("implicitDeps[{}]", index);
+
+        // When no target scope, it's assumed to be a self scope
+        if item.contains(':') {
+            validate_target(key, item)?;
+        } else {
+            validate_id(key, item)?;
+        }
+    }
+
+    Ok(())
+}
 
 fn validate_file_groups(map: &FileGroups) -> Result<(), ValidationError> {
     for key in map.keys() {
@@ -56,6 +71,13 @@ pub struct InheritedTasksConfig {
     pub file_groups: FileGroups,
 
     #[serde(skip_serializing_if = "is_default")]
+    #[validate(custom = "validate_deps")]
+    pub implicit_deps: Vec<String>,
+
+    #[serde(skip_serializing_if = "is_default")]
+    pub implicit_inputs: Vec<String>,
+
+    #[serde(skip_serializing_if = "is_default")]
     #[validate(custom = "validate_tasks")]
     #[validate]
     pub tasks: BTreeMap<String, TaskConfig>,
@@ -74,16 +96,28 @@ impl InheritedTasksConfig {
             let figment = Figment::from(YamlExtended::file(source).profile(profile_name));
             let extended_config = InheritedTasksConfig::load_config(figment.select(profile_name))?;
 
-            // Figment does not merge hash maps but replaces entirely,
+            // Figment does not merge maps/vec but replaces entirely,
             // so we need to manually handle this here!
             if !extended_config.file_groups.is_empty() {
                 config.file_groups.extend(extended_config.file_groups);
+            }
+
+            if !extended_config.implicit_deps.is_empty() {
+                config.implicit_deps.extend(extended_config.implicit_deps);
+            }
+
+            if !extended_config.implicit_inputs.is_empty() {
+                config
+                    .implicit_inputs
+                    .extend(extended_config.implicit_inputs);
             }
 
             if !extended_config.tasks.is_empty() {
                 config.tasks.extend(extended_config.tasks);
             }
         }
+
+        config.implicit_inputs.push("/.moon/*.yml".into());
 
         Ok(config)
     }
